refactor: moving the ActorTile, Error Page, Image PlaceHolder, Loading Page, MovieTile, NoResultsPage and Tile components to the common folder
