+++ src/common/ImagePlaceholder/styled.js
