+++ src/common/Tile/styled.js
